mvp done: use token instead of user-id, endpoint to get usage stats by app

+++ wayd-backend/src/main/java/blck/wayd/data/entity/TrackLog.java
@@ -26,7 +26,7 @@ public class TrackLog {
     private String appName;
 
     @NonNull
-    private Integer timestamp;
+    private Long timestamp;
 
     @NonNull
     private UUID userId;
